* LCIMClient.cs: * Realtime.csproj: * Program.cs: * LCIMMessage.cs: * Conversation.cs: * LCIMChatRoom.cs: * LCRTMRouter.cs: * LCIMConversation.cs: * LCIMConversationQuery.cs: * LCIMController.cs: * LCIMGoAwayController.cs: * LCIMUnreadController.cs: * LCWebSocketConnection.cs: * LCIMMessageController.cs: * LCIMSessionController.cs: * LCIMConversationController.cs:
* LCIMTypedMessage.cs: refactor

+++ Realtime/Message/LCIMTypedMessage.cs
@@ -55,7 +55,7 @@ namespace LeanCloud.Realtime {
 
         internal static LCIMTypedMessage Deserialize(Dictionary<string, object> messageData) {
             LCIMTypedMessage message = null;
-            int msgType = (int)(long)messageData["_lctype"];
+            int msgType = (int)messageData["_lctype"];
             switch (msgType) {
                 case -1:
                     message = new LCIMTextMessage();
